Safari-609.1.11.1/LayoutTests/platform/wpe/fast/canvas/webgl/webgl-draw-buffers-expected.txt

CONSOLE MESSAGE: line 205: WebGL: INVALID_ENUM: getParameter: invalid parameter name, WEBGL_draw_buffers not enabled
CONSOLE MESSAGE: line 205: WebGL: INVALID_ENUM: getParameter: invalid parameter name, WEBGL_draw_buffers not enabled
CONSOLE MESSAGE: line 1313: WebGL: ERROR: 0:2: 'GL_EXT_draw_buffers' : extension is not supported
CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:2: 'GL_EXT_draw_buffers' : extension is not supported

CONSOLE MESSAGE: line 24: Error in compiling shader
CONSOLE MESSAGE: line 294: WebGL: INVALID_ENUM: framebufferTexture2D: invalid attachment
This test verifies the functionality of the WEBGL_draw_buffers extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists

Testing binding enum with extension disabled
PASS getError was expected value: INVALID_ENUM : MAX_COLOR_ATTACHMENTS_WEBGL should not be queryable if extension is disabled
PASS getError was expected value: INVALID_ENUM : MAX_DRAW_BUFFERS_WEBGL should not be queryable if extension is disabled
PASS getError was expected value: NO_ERROR : there should be no errors

test shaders disabled
vertex shader
fragement shader
PASS GL_EXT_draw_buffers should not be defined in GLSL
vertex shader
fragement shader
PASS #extension GL_EXT_draw_buffers should not be allowed in GLSL
PASS getError was expected value: NO_ERROR : there should be no errors

test attachment disabled
PASS getError was expected value: INVALID_ENUM : should not be able to attach to gl.COLOR_ATTACHMENT1
PASS getError was expected value: NO_ERROR : there should be no errors

PASS No WEBGL_draw_buffers support -- this is legal
PASS WEBGL_draw_buffers not listed as supported and getExtension failed -- this is legal

PASS successfullyParsed is true

TEST COMPLETE
